Memoize parsing View definitions into Table objects
This patch drops CPU time by about 20x, since profiling
showed that about 50% of time is spent in parsing SQL. We
could also make the parser itself faster, but this seemed to be
a simpler solution.

Also modififes the regex cleaning routine to clean up the name
of the databases as well in the 'from' clause, to increase
hit rate on the memoization. It will be 0 otherwise.

Also stops setting .table on column, to make the data structure
acyclic.

Change-Id: Ibf4f91b07ea45029561ab399b6aea6f00db82aaa